+++ just_audio/CHANGELOG.md
@@ -1,3 +1,20 @@
+## 0.10.1
+
+* 
+
+## 0.9.24
+
+* Bump Android compileSdkVersion to 31.
+
+## 0.9.23
+
+* Fallback to constant bitrate seeking on Android.
+
+## 0.9.22
+
+* Migrate to Flutter 3, Xcode 13, ExoPlayer 2.17.1.
+* Add just_audio_mpv and just_audio_windows to example.
+
 ## 0.9.21
 
 * Handle and report network disconnections in StreamAudioSource.